% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_units}
\alias{add_units}
\title{Add EML units to the attribute table}
\usage{
add_units(data_list)
}
\arguments{
\item{data_list}{A large list from
\code{\link[fetchsqlserver]{fetch_sql_server_data()}}.}
}
\value{
A large list (data_list) with the updated attribute table(s).
}
\description{
Use this function after \code{\link{convert_data_type_to_class()}} to add EML
 units to the 'numeric' classes in all the attribute tables. A dictionary is
used to look up the attributeName
and get the EML unit. If an attributeName(s) is not found in the dictionary,
the attributeName(s) with missing EML units will be provided to the user.
Also, if a unit is not a standard EML unit, a TSV file with the custom units,
 custom_units.txt, will be saved to the working directory. This file is
needed to create EML metadata.
}
