% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{convert_data_type_to_class}
\alias{convert_data_type_to_class}
\title{Convert attribute table data types to EML classes}
\usage{
convert_data_type_to_class(attr_df)
}
\arguments{
\item{attr_df}{A data frame of the attribute table created by
\code{\link{fetch_attr_tables()}}}
}
\value{
A data frame of the attribute table with the EML classes, and
dateTimeFormatString added, if applicable.
}
\description{
When the attribute tables are fetched using
\code{\link{fetch_attr_tables()}}, the SQL Server data types for
each column are are stored under the 'unit' column. These data types need
to be converted to the appropriate EML class to produce EML metadata
attribute tables for the data package. \code{convert_data_type_to_class()}
will do this conversion. It will also add the applicable dateTimeFormatString
 to the date classes.
}
