% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{create_catvar_tbls}
\alias{create_catvar_tbls}
\title{Create categorical variable tables for applicable data tables to use
in EML metadata creation.}
\usage{
create_catvar_tbls(data_list)
}
\arguments{
\item{data_list}{A large list from \code{\link{fetch_sql_server_data()}}}
}
\value{
A large list containing [[1]] the data_list with the updated
attribute tables, [[2]] a vector with all the table names (strings) that
have catvar tables, and [[3]] a list of catvar data frames. Before
running \code{\link{save_data_pkg()}}, the data_list and catvar table names
list and catvar data frames list need to be extracted (see example).
}
\description{
Using the RefSchemaName and RefTableName columns from the attribute table,
along with catvar_tables.csv
(created by \code{\link{fetch_sql_server_data()}}
or \code{\link{fetch_attr_tables()}} and filled out by the user), this
function creates a catvar data
frames with attributeName, code, and definition columns Also, updates the
attribute table class to 'categorical' for all columns with catvars. The
catvar_tables.csv file must be populated before running this function, and it
must have the original file name.
}
