% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEML.R
\name{create_eml_metadata}
\alias{create_eml_metadata}
\title{Create EML metadata (non-NPS-specific portion)}
\usage{
create_eml_metadata(eml_info)
}
\arguments{
\item{eml_info}{A hash dictionary created by
\code{\link{gather_eml_info()}} containing the base metadata needed
 to create EML.}

\item{data_pkg_folder}{A string of the folder path for the data package
. If \code{save_data_pkg()} was run before running
\code{save_core_metadata}, then the data package folder path should be
saved as a global variable.}
}
\value{
A list containing metadata information such as overall package
details, geographic, taxonomic and temporal information, and core metadata
information. Also called an \code{eml_object} by
\code{\link[EMLeditor]{EMLeditor}}. Use as an input to
\code{\link{add_nps_metadata()}}.
}
\description{
Creates the base EML metadata sans the NPS-specific stuff, which is added
using \code{\link{add_nps_metadata()}}. Ensures that all core metadata files
are present prior to creating an \code{eml_object}.
}
