% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNPSmetadata.R
\name{create_nps_eml}
\alias{create_nps_eml}
\title{Add NPS-specific fields to base EML}
\usage{
create_nps_eml(eml_info, my_metadata)
}
\arguments{
\item{eml_info}{A hash dictionary created by
\code{\link{gather_eml_info()}} and \code{\link{gather_nps_info()}}
 containing the metadata needed to create EML.}

\item{my_metadata}{A list containing metadata information such as overall
package details, geographic, taxonomic and temporal information, and core
metadata information. Also called an \code{eml_object} by
\code{\link[EMLeditor]{EMLeditor}}.}
}
\value{
A list containing metadata information such as overall
package details, geographic, taxonomic and temporal information, core
metadata information, and  NPS-specific metadata. Also called an
\code{eml_object} by \code{EMLeditor}.
}
\description{
After the base EML has been created using \code{\link{create_eml_metadata()}}
, this function is used to add the NPS-specific fields. The metadata can then
 be written to XML using \code{\link{write_eml_xml()}}.
}
