% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_attr_tables}
\alias{fetch_attr_tables}
\title{Fetch attributes and foreign key references using queries}
\usage{
fetch_attr_tables(qry_path = NULL)
}
\arguments{
\item{qry_path}{(optional) A string with the folder path to the query files
for fetching the data tables. If \code{NULL}, the queries will be pulled
from the folder named "queries" in the working directory. The working
directory is set when \code{\link{read_data_pkg_info()}} is ran. Or you can
set the working directory yourself using using \code{setwd()} or
\code{\link{set_pkg_dir()}} However,
if the "queries" folder is not found in the working directory, the user is
prompted to select the folder.}
}
\value{
A list of data frames containing the table attributes.
}
\description{
Get the attributes for the data tables, as well as the reference schema and
table for any foreign keys (RefSchemaName and RefTableName). The resulting
data frame column names are formatted for use with EMLeditor: attributeName
(column name), attributeDefinition (column description), class (data type),
unit (empty), dataTimeStringFormat (empty), missingValueCode (empty), and
missingValueCodeExplanation (empty).

Additionally, two CSVs are saved for the user to fill out. catvar_tables.csv
is created using the foreign keys associated with data table attributes. The
CSV contains a list of lookup/ref tables with their schema and table name, as
 well as a Code and Description column. The Code column where the user fills
in the column or field name from the lookup/ref table associated with the
data table. The Description column is where the user fills in the column or
or field from the lookup/ref table that contains the description of the Code.

This function also converts the attribute (field) data types to EML classes.
All attributes with a numeric field must have a unit specified in the
attribute table. field_unit_dict.csv contains a list of all the fields with
a numeric class. Under the Unit column, the user fills out the unit for that
field. The unit must be from the EMLassemblyline unit dictionary. This
dictionary can be found by running
\code{\link[EMLassemblyline]view_unit_dictionary()}, and it is also saved in
the data package info Excel spreadsheet under the units_lists sheet.

Not meant to be used on its own; it is primarily meant to only be used in
\code{\link{fetch_sql_server_data}}.
}
