% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_data_tables}
\alias{fetch_data_tables}
\title{Fetch data tables}
\usage{
fetch_data_tables(qry_path)
}
\arguments{
\item{qry_path}{A string with the folder path to the query files
for fetching the data tables. If \code{NULL}, the queries will be pulled
from the folder named "queries" in the working directory. The working
directory is set when \code{\link{read_data_pkg_info()}} is ran. Or you can
set the working directory yourself using using \code{setwd()} or
\code{\link{set_pkg_dir()}}. However,
if the "queries" folder is not found in the working directory, the user is
prompted to select the folder.}
}
\value{
A large list of data frames.
}
\description{
Fetches data tables using queries stored in .sql or .txt files. The queries
must be in a specific format. See documentation for more details.

Not meant to be used on its own; it is primarily meant to only be used in
\code{\link{fetch_sql_server_data}}.
}
