% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_sql_server_data}
\alias{fetch_sql_server_data}
\title{Fetch SQL Server data}
\usage{
fetch_sql_server_data(qry_path = NULL)
}
\arguments{
\item{qry_path}{(optional) A string with the folder path to the query files
for fetching the data tables. If \code{NULL}, the queries will be pulled
from the folder named "queries" in the working directory. The working
directory is set when \code{\link{read_data_pkg_info()}} is ran. Or you can
set the working directory yourself using using \code{setwd()} or
\code{\link{set_pkg_dir()}} However,
if the "queries" folder is not found in the working directory, the user is
prompted to select the folder.}
}
\value{
A large list consisting of [[1]]
data tables in data frames from 'data' schema, [[2]] attributes and
foreign keys for those data tables in data frames, and [[3]] a data frame
with table names and descriptions.
}
\description{
Fetches the data tables using the user-provided queries, the data tables'
attributes and foreign keys, and data table descriptions. For
more information on the format of the attributes and foreign keys, see
\code{\link{fetch_attr_tbls_fk}}. The data table descriptions come from the
 database extended properties. The resulting data frame is formatted to be
used with functions from \code{fetchsqlserver} (see documentation).
}
