% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEML.R
\name{gather_eml_info}
\alias{gather_eml_info}
\title{Gather base EML metadata info (non-NPS-specific portion)}
\usage{
gather_eml_info(data_pkg_folder = NULL)
}
\arguments{
\item{data_pkg_folder}{(optional) A string of the folder path for the data
package.}
}
\value{
A hash dictionary containing the base metadata needed to create EML.
}
\description{
Gathers overall package details, taxonomic, geographic, and temporal
information by reading files and asking for user input.
data_names_descriptions.xlsx must be present in the data package folder,
which is added using \code{\link{save_data_pkg()}}. After running
\code{gather_eml_info()}, the next step is to run
\code{\link{create_eml_metadata()}}.
}
