% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNPSmetadata.R
\name{gather_nps_info}
\alias{gather_nps_info}
\title{Gather NPS-specific metadata information}
\usage{
gather_nps_info(eml_info)
}
\arguments{
\item{eml_info}{A hash dictionary created by
\code{\link{gather_eml_info()}} containing the base metadata needed
 to create EML.}
}
\value{
A hash dictionary created by containing the base and NPS-specific
metadata needed to create EML.
}
\description{
Gathers the NPS-specific metadata information such as intellectual
rights, CUI code, park units, producing units, and DataStore reference using
 user inputs. After running \code{gather_nps_info()}, the next step is to run
\code{\link{add_nps_metadata()}}.
}
