% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataListTools.R
\name{get_tbl_as_df}
\alias{get_tbl_as_df}
\title{Get table from data_list as a data frame}
\usage{
get_tbl_as_df(table_name, table_type = "data", data_list)
}
\arguments{
\item{table_name}{A string of the table name associated with the data table
data frame or attribute table data frame as it appears in the dataFile
column of the table names and descriptions data frame (data_list[[3]]).}

\item{table_type}{A string with the type of table you wish to retrieve from
the data_list, either \code{'data'} for a data table or \code{'attr'} for an
attribute table. The default is \code{'data'}.}

\item{data_list}{A large list from
\code{\link[fetchsqlserver]{fetch_sql_server_data}}.}
}
\value{
A data frame with either the specified data or attribute table
pulled from the large list (data_list).
}
\description{
Gets either a data table or attribute table from the large list returned from
 \code{\link[fetchsqlserver]{fetch_sql_server_data}}. This allows you to
work with either or both tables more easily than working with them while they
 are in the large list.
}
\examples{
\dontrun{
# with a list of table names
tbl_list <- c(
  "ChannelConstraints",
  "ConstrainingFeatures",
  "EventConditions",
  "EventLog",
  "Events",
  "HabitatAssessment",
  "HumanInfluences",
  "InvasiveObs",
  "Locations",
  "TorrentEvidence",
  "WidthDepth",
  "WQData",
  "WQSamples"
)

db_name <- "WCa01_Water_Quality"

data_list <- fetch_sql_server_data(db_name, tbl_list)

# Let's say you pulled the WQData data table and attribute table from the
# the data_list to do some wrangling or modifications.
wqdata_df <- get_tbl_as_df("WQData", data_list)
wqdata_attr_df <- get_tbl_as_df("WQData", "attr", data_list)

# Pulling out the data frames makes them easier to work (in Christina's
# opinion) with because they have a (variable) name...
wqdata_df \%<>\% select(-ProcessingFlag)

# ...instead of index numbers
data_list[[2]][[12]] \%<>\% select(-ProcessingFlag)

# But do whatever feels right to you!
}

}
