% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{missing_values}
\alias{missing_values}
\title{Add missing value info to attribute table or get a list of columns
with missing data}
\usage{
missing_values(data_df, attr_df = NULL)
}
\arguments{
\item{data_df}{A data table data frame pulled from the large list (returned
from \code{\link{fetch_sql_server_data()}}) using
\code{\link{get_tbl_as_df()}} or another means.}

\item{attr_df}{(optional) A data frame of the attribute table pulled from
\code{\link{fetch_sql_server_data()}})
using \code{\link{get_tbl_as_df()}} or another means. The default is
\code{NULL}, and if this parameter is \code{NULL}, only a vector of column
names that contain \code{NA} or \code{NULL} values is returned.}
}
\value{
An attribute table data frame, if one was provided, with updated
missingValueCode and missingValueCodeExplanations for \code{NA} or
\code{NULL} columns, otherwise, a vector of column names that contain
\code{NA} or \code{NULL} values.
}
\description{
Adds generic missing value information to the attribute table if a column
contains NA or null values, or gets a list of columns with \code{NA} or
\code{NULL} values if the missing value info in the attribute table needs to
be customized beyond the generic missingValueCode and
missingValueCodeExplanation provided in the data pkg info Excel spreadsheet.

\code{missing_values} checks to see if the missingValueCode is \code{NA} or
\code{NULL}. If it is, it is assumed that no custom values have been entered
for missingValueCode AND missingValueCodeExplanation. As a result, this
function can be used after customizing applicable missing value information
to fill in the rest with the generic info.
}
\seealso{
\code{\link{missing_values_all_tbls()}}
}
