% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_cust_units_sheet}
\alias{read_cust_units_sheet}
\title{Read custom units sheet in data pkg info Excel Spreadsheet}
\usage{
read_cust_units_sheet(data_pkg_info_path = NULL)
}
\description{
Once the data pkg info spreadsheet has been filled out by the user (see
\code{\link{save_templates()}}), the function
\code{\link{read_data_pkg_info()}} should be ran to read the contents. If
changes were made to the custom units after running that function, this
function needs to be ran to update the custom units. This function reads the
XLSX file and saves the custom units in a hidden data frame. The data frame
is used by \code{\link{add_units()}}.
}
