% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data_pkg_info}
\alias{read_data_pkg_info}
\title{Read data pkg info Excel spreadsheet}
\usage{
read_data_pkg_info()
}
\description{
Once the data pkg info spreadsheet has been filled out by the user (see
\code{\link{save_data_info_template()}}), this function reads both sheets in
the spreadsheet and saves the information in hidden hash dictionaries and
data frames. If either of the sheets are not saved/set, a message will appear
 prompting the user to run the function for that individual sheet.
}
