% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_overall_sheet}
\alias{read_overall_sheet}
\title{Read overall sheet in data pkg info Excel Spreadsheet}
\usage{
read_overall_sheet(data_pkg_info_path = NULL)
}
\description{
Once the data pkg info spreadsheet has been filled out by the user (see
\code{\link{save_templates()}}), the function
\code{\link{read_data_pkg_info()}} should be ran to read the contents. If
changes were made to the overall info after running that function, this
function needs to be ran to update the overall info.
This function reads the XLSX file
and saves the overall information in a hidden hash dictionary. The overall
information is used by many functions in the \code{fetchsqlserver} package.
}
