% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataListTools.R
\name{rearrange_attr_tbl}
\alias{rearrange_attr_tbl}
\title{Rearrange attribute table to match data table}
\usage{
rearrange_attr_tbl(data_df, attr_df)
}
\arguments{
\item{data_df}{A data table data frame extracted from the data_list using
\code{\link{get_tbl_as_df}} that has been modified.}

\item{attr_df}{An attribute table data frame extracted from the data_list
using \code{\link{get_tbl_as_df}}.}
}
\value{
An attribute table data frame with attributeNames matching the input
data table data frame.
}
\description{
After deleting and/or rearranging columns in the data table,
\code{rearrange_attr_tbl} will update the attribute table to have the same
columns in the same order as the data table.
}
