% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataListTools.R
\name{remove_from_data_list}
\alias{remove_from_data_list}
\title{Remove table from data list}
\usage{
remove_from_data_list(table_name, data_list)
}
\arguments{
\item{table_name}{A string of the table name associated with the data table
data frame or attribute table data frame as it appears in the dataFile
column of the table names and descriptions data frame (data_list[[3]]).}

\item{data_list}{A large list from
\code{\link[fetchsqlserver]{fetch_sql_server_data}}.}
}
\value{
A large list (data_list) with the specified table data removed.
}
\description{
Use on the large list returned from
\code{\link[fetchsqlserver]{fetch_sql_server_data}}
. If a data table and it's associated attribute table and table description
are no longer needed and will not be published with the data package, this
function will remove the data table
(data_list[[1]]), attribute table (data_list[[2]]), and table
description (data_list[[3]]) from the large list.
}
\examples{
\dontrun{
# with a list of table names
tbl_list <- c(
  "ChannelConstraints",
  "ConstrainingFeatures",
  "EventConditions",
  "EventLog",
  "Events",
  "HabitatAssessment",
  "HumanInfluences",
  "InvasiveObs",
  "Locations",
  "TorrentEvidence",
  "WidthDepth",
  "WQData",
  "WQSamples"
)

db_name <- "WCa01_Water_Quality"

data_list <- fetch_sql_server_data(db_name, tbl_list)

# Let's say the Events and Locations tables were merged together and Events
#  is no longer needed.
data_list <- remove_from_data_list("Events", data_list)
}

}
