% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_data_info_template}
\alias{save_data_info_template}
\title{Save data pkg info Excel spreadsheet template}
\usage{
save_data_info_template()
}
\description{
Saves the data pkg info spreadsheet that is to be
filled out by the user containing information needed by many functions in the
 \code{fetchsqlserver} package. The first sheet, overall, is for information
such as server and database names, park units, date format, etc. The second
sheet, custom_units, is for any custom units that are used in the
attribute tables for metadata creation. The template already contains four
custom units, which can be deleted at the user's discretion. These are
provided as an example of how to fill out the sheet. Any units not
found in the EMLassemblyline unit dictionary (also available in the third
sheet, units_list) must be added to custom_units_table.csv.
}
