% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{save_data_pkg}
\alias{save_data_pkg}
\title{Save data package files}
\usage{
save_data_pkg(data_list, catvar_list)
}
\arguments{
\item{data_list}{A large list from
\code{\link[fetchsqlserver]{fetch_sql_server_data()}}.}

\item{catvar_list}{A large list from \code{\link{create_catvar_tbls()}}.}
}
\description{
Saves the data tables to CSVs, the attribute and catvar tables to TSVs, and
saves the data names and descriptions to an XLSX (used to create EML metadata
) using \code{\link{create_eml_metadata()}}. When \code{save_data_pkg()}
runs, the folder path selected for the data package folder will be saved as a
 global variable. If continuing with metadata creation, this global variable
will be used to read/write to the data package folder. The following should
be complete before running \code{save_data_pkg()}:
* Data tables are wrangled, flattened, cleaned, etc.
* Attribute data types converted to EML class using
\code{\link{convert_data_type_class()}}, if necessary
* Units added to attribute tables using
\code{\link{add_units()}}
* Missing value info added to attribute tables using
\code{\link{missing_values()}} and/or \code{\link{missing_values_all_tbls()}}
* Categorical variable tables created using
\code{\link{create_catvar_tbls()}}
}
