% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_pkg_wd}
\alias{set_pkg_wd}
\title{Set package working directory}
\usage{
set_pkg_wd()
}
\description{
This function does the same thing as \code{setwd()}, but it allows the user
to use the File Explorer to select the directory instead of typing out the
folder path or copying and pasting. That way the user doesn't have to worry
about which slashes they use (forward or back?), how many slashes, or
mistyping the path.
}
