% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataListTools.R
\name{update_data_list}
\alias{update_data_list}
\title{Update data list data frame}
\usage{
update_data_list(table_name, data_list, data_df = NULL, attr_df = NULL)
}
\arguments{
\item{table_name}{A string of the table name associated with the data table
data frame or attribute table data frame as it appears in the dataFile
column of the table names and descriptions data frame (data_list[[3]]).}

\item{data_list}{A large list from
\code{\link[fetchsqlserver]{fetch_sql_server_data}}.}

\item{data_df}{A data table data frame extracted from the data_list using
\code{\link{get_tbl_as_df}} that has been modified and is to replace the
data table data frame in the data_list.}

\item{attr_df}{An attribute table data frame extracted from the data_list
using \code{\link{get_tbl_as_df}} that has been modified and is to replace
the attribute table data frame in the data_list.}
}
\value{
A large list (data_list) with the updated data frame(s).
}
\description{
Use on the large list returned from
\code{\link[fetchsqlserver]{fetch_sql_server_data}}
. After a data table or attribute table has been pulled from the large list
and modified, this function replaces the table(s) with the modified tables.
 There is no need to use this function if you work with the data or attribute
 table in the large list.
}
\examples{
\dontrun{
# with a list of table names
tbl_list <- c(
  "ChannelConstraints",
  "ConstrainingFeatures",
  "EventConditions",
  "EventLog",
  "Events",
  "HabitatAssessment",
  "HumanInfluences",
  "InvasiveObs",
  "Locations",
  "TorrentEvidence",
  "WidthDepth",
  "WQData",
  "WQSamples"
)

db_name <- "WCa01_Water_Quality"

data_list <- fetch_sql_server_data(db_name, tbl_list)

# Let's say you pulled the WQData data table and attribute table from the
# the data_list to do some wrangling or modifications.
wqdata_df <- get_tbl_as_df("WQData", table_type = "data", data_list)
wqdata_attr_df <- get_tbl_as_df("WQData", table_type = "attr", data_list)
wqdata_df <- get_tbl_as_df("WQData", table_type = "data", data_list)
wqdata_attr_df <- get_tbl_as_df("WQData", table_type = "attr", data_list)

# After the wrangling or modification are complete, replace the data frames
# in the data_list

data_list <- update_data_list("WQData", data_list, wqdata_df, wqdata_attr_df)
}

}
