% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNPSmetadata.R
\name{write_eml_xml}
\alias{write_eml_xml}
\title{Write EML metadata to XML file}
\usage{
write_eml_xml(my_metadata, eml_info = NULL)
}
\arguments{
\item{my_metadata}{A list containing metadata information such as overall
package details, geographic, taxonomic and temporal information, core
metadata information, and  NPS-specific metadata. Also called an
\code{eml_object} by \code{\link[EMLeditor]{EMLeditor}}.}

\item{eml_info}{(optional) A hash dictionary created by containing the base
and NPS-specific metadata needed to create EML generated by
\code{\link{gather_eml_info()}} and \code{\link{gather_nps_info()}}.}
}
\value{
A hash dictionary created by containing the base
and NPS-specific metadata
}
\description{
Writes EML metadata saved as an \code{eml_object} to an XML file. If an
\code{eml_info} object is passed in, the \code{eml_info} object will be
updated with the DataStore reference. This can be useful if data package
creation is an iterative process.
}
